Renamed InvocationValue to InvocationConstant
The InvocationConstant contains the constants
we use to determine all kinds of internal state
for the invocation/invocationfuture.
